requirements.txt
streamlit
openai
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-api-python-client
